+++ src/SimulateTurn/Actions/ResolveDecision.cpp
@@ -14,7 +14,7 @@
 #include <Types/Entity.hpp>
 #include <Types/Enums/ActionOrder.hpp>
 #include <Types/State.hpp>
-#include <Types/Utilities/variant.hpp>
+#include <Types/Utilities/Variant.hpp>
 #include <entt/entity/handle.hpp>
 #include <entt/entity/registry.hpp>
